+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Support parameterised types other than `Array` and `Object`, such as `Foo<T>`
 
-## [0.3.5] - 2017-01-01
+## [0.3.6] - 2017-01-09
+- Support parameterized types other than `Array` and `Object`, such as `Foo<T>`.
+
+## [0.3.5] - 2017-01-02
 - Properties are now emitted as `readonly` when applicable.
 - Bump Analyzer for latest scanning features (getters/setters, static methods, methods/properties on class prototypes).